+++ src/main/java/fr/istic/web/rest/ExamSheetResource.java
@@ -2,8 +2,12 @@ package fr.istic.web.rest;
 
 import static javax.ws.rs.core.UriBuilder.fromPath;
 
+import fr.istic.domain.Authority;
+import fr.istic.domain.ExamSheet;
+import fr.istic.domain.User;
 import fr.istic.security.AuthoritiesConstants;
 import fr.istic.service.ExamSheetService;
+import fr.istic.web.rest.errors.AccountResourceException;
 import fr.istic.web.rest.errors.BadRequestAlertException;
 import fr.istic.web.util.HeaderUtil;
 import fr.istic.web.util.ResponseUtil;
@@ -13,6 +17,7 @@ import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import fr.istic.service.Paged;
+import fr.istic.service.SecurityService;
 import fr.istic.web.rest.vm.PageRequestVM;
 import fr.istic.web.rest.vm.SortRequestVM;
 import fr.istic.web.util.PaginationUtil;
@@ -42,17 +47,22 @@ public class ExamSheetResource {
     @ConfigProperty(name = "application.name")
     String applicationName;
 
-
     @Inject
     ExamSheetService examSheetService;
+
+    @Inject
+    SecurityService securityService;
+
     /**
      * {@code POST  /exam-sheets} : Create a new examSheet.
      *
      * @param examSheetDTO the examSheetDTO to create.
-     * @return the {@link Response} with status {@code 201 (Created)} and with body the new examSheetDTO, or with status {@code 400 (Bad Request)} if the examSheet has already an ID.
+     * @return the {@link Response} with status {@code 201 (Created)} and with body
+     *         the new examSheetDTO, or with status {@code 400 (Bad Request)} if the
+     *         examSheet has already an ID.
      */
     @POST
-    @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
+    @RolesAllowed({ AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN })
     public Response createExamSheet(@Valid ExamSheetDTO examSheetDTO, @Context UriInfo uriInfo) {
         log.debug("REST request to save ExamSheet : {}", examSheetDTO);
         if (examSheetDTO.id != null) {
@@ -60,7 +70,8 @@ public class ExamSheetResource {
         }
         var result = examSheetService.persistOrUpdate(examSheetDTO);
         var response = Response.created(fromPath(uriInfo.getPath()).path(result.id.toString()).build()).entity(result);
-        HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, result.id.toString()).forEach(response::header);
+        HeaderUtil.createEntityCreationAlert(applicationName, true, ENTITY_NAME, result.id.toString())
+                .forEach(response::header);
         return response.build();
     }
 
@@ -68,20 +79,28 @@ public class ExamSheetResource {
      * {@code PUT  /exam-sheets} : Updates an existing examSheet.
      *
      * @param examSheetDTO the examSheetDTO to update.
-     * @return the {@link Response} with status {@code 200 (OK)} and with body the updated examSheetDTO,
-     * or with status {@code 400 (Bad Request)} if the examSheetDTO is not valid,
-     * or with status {@code 500 (Internal Server Error)} if the examSheetDTO couldn't be updated.
+     * @return the {@link Response} with status {@code 200 (OK)} and with body the
+     *         updated examSheetDTO,
+     *         or with status {@code 400 (Bad Request)} if the examSheetDTO is not
+     *         valid,
+     *         or with status {@code 500 (Internal Server Error)} if the
+     *         examSheetDTO couldn't be updated.
      */
     @PUT
-    @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response updateExamSheet(@Valid ExamSheetDTO examSheetDTO) {
+    @RolesAllowed({ AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN })
+    public Response updateExamSheet(@Valid ExamSheetDTO examSheetDTO, @Context SecurityContext ctx) {
         log.debug("REST request to update ExamSheet : {}", examSheetDTO);
         if (examSheetDTO.id == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+        if (!securityService.canAccess(ctx, examSheetDTO.id, ExamSheet.class)) {
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
+
         var result = examSheetService.persistOrUpdate(examSheetDTO);
         var response = Response.ok().entity(result);
-        HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, examSheetDTO.id.toString()).forEach(response::header);
+        HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, examSheetDTO.id.toString())
+                .forEach(response::header);
         return response.build();
     }
 
@@ -93,12 +112,18 @@ public class ExamSheetResource {
      */
     @DELETE
     @Path("/{id}")
-    @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response deleteExamSheet(@PathParam("id") Long id) {
+    @RolesAllowed({ AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN })
+    public Response deleteExamSheet(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to delete ExamSheet : {}", id);
+        if (!securityService.canAccess(ctx, id, ExamSheet.class)) {
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
+        ;
+
         examSheetService.delete(id);
         var response = Response.noContent();
-        HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString()).forEach(response::header);
+        HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString())
+                .forEach(response::header);
         return response.build();
     }
 
@@ -106,43 +131,61 @@ public class ExamSheetResource {
      * {@code GET  /exam-sheets} : get all the examSheets.
      *
      * @param pageRequest the pagination information.
-     * @return the {@link Response} with status {@code 200 (OK)} and the list of examSheets in body.
+     * @return the {@link Response} with status {@code 200 (OK)} and the list of
+     *         examSheets in body.
      */
     @GET
-    public Response getAllExamSheets(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo) {
+    public Response getAllExamSheets(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest,
+            @Context UriInfo uriInfo, @Context SecurityContext ctx) {
         log.debug("REST request to get a page of ExamSheets");
         var page = pageRequest.toPage();
         var sort = sortRequest.toSort();
         Paged<ExamSheetDTO> result = null;
         MultivaluedMap param = uriInfo.getQueryParameters();
 
-        if (param.containsKey("name") ) {
+        if (param.containsKey("name")) {
             List name = (List) param.get("name");
-            result = examSheetService.findExamSheetByName(page, ""+name.get(0));
-        }else {
-            result =examSheetService.findAll(page);
+            result = examSheetService.findExamSheetByName(page, "" + name.get(0));
+        } else {
 
+            var userLogin = Optional
+                    .ofNullable(ctx.getUserPrincipal().getName());
+            if (!userLogin.isPresent()) {
+                throw new AccountResourceException("Current user login not found");
+            }
+            var user = User.findOneByLogin(userLogin.get());
+            if (!user.isPresent()) {
+                throw new AccountResourceException("User could not be found");
+            } else if (user.get().authorities.size() >= 1
+                    && user.get().authorities.stream().anyMatch(e1 -> e1.equals(new Authority("ROLE_ADMIN")))) {
+                result = examSheetService.findAll(page);
+
+            } else {
+                return Response.status(403, "Current user cannot access to this ressource").build();
+            }
         }
-
-
-
         var response = Response.ok().entity(result.content);
         response = PaginationUtil.withPaginationInfo(response, uriInfo, result);
         return response.build();
     }
 
-
     /**
      * {@code GET  /exam-sheets/:id} : get the "id" examSheet.
      *
      * @param id the id of the examSheetDTO to retrieve.
-     * @return the {@link Response} with status {@code 200 (OK)} and with body the examSheetDTO, or with status {@code 404 (Not Found)}.
+     * @return the {@link Response} with status {@code 200 (OK)} and with body the
+     *         examSheetDTO, or with status {@code 404 (Not Found)}.
      */
     @GET
     @Path("/{id}")
-
-    public Response getExamSheet(@PathParam("id") Long id) {
+    @RolesAllowed({ AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN })
+    public Response getExamSheet(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to get ExamSheet : {}", id);
+        if (!securityService.canAccess(ctx, id, ExamSheet.class)) {
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
+        ;
+
         Optional<ExamSheetDTO> examSheetDTO = examSheetService.findOne(id);
         return ResponseUtil.wrapOrNotFound(examSheetDTO);
     }
