+++ src/main/java/fr/istic/web/rest/GradedCommentResource.java
@@ -2,8 +2,13 @@ package fr.istic.web.rest;
 
 import static javax.ws.rs.core.UriBuilder.fromPath;
 
+import fr.istic.domain.Authority;
+import fr.istic.domain.GradedComment;
+import fr.istic.domain.Question;
+import fr.istic.domain.User;
 import fr.istic.security.AuthoritiesConstants;
 import fr.istic.service.GradedCommentService;
+import fr.istic.web.rest.errors.AccountResourceException;
 import fr.istic.web.rest.errors.BadRequestAlertException;
 import fr.istic.web.util.HeaderUtil;
 import fr.istic.web.util.ResponseUtil;
@@ -13,6 +18,7 @@ import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import fr.istic.service.Paged;
+import fr.istic.service.SecurityService;
 import fr.istic.web.rest.vm.PageRequestVM;
 import fr.istic.web.rest.vm.SortRequestVM;
 import fr.istic.web.util.PaginationUtil;
@@ -44,6 +50,10 @@ public class GradedCommentResource {
 
     @Inject
     GradedCommentService gradedCommentService;
+
+    @Inject
+    SecurityService securityService;
+
     /**
      * {@code POST  /graded-comments} : Create a new gradedComment.
      *
@@ -73,10 +83,13 @@ public class GradedCommentResource {
      */
     @PUT
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response updateGradedComment(GradedCommentDTO gradedCommentDTO) {
+    public Response updateGradedComment(GradedCommentDTO gradedCommentDTO, @Context SecurityContext ctx) {
         log.debug("REST request to update GradedComment : {}", gradedCommentDTO);
         if (gradedCommentDTO.id == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
+        }
+        if (!securityService.canAccess(ctx, gradedCommentDTO.id, GradedComment.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
         }
         var result = gradedCommentService.persistOrUpdate(gradedCommentDTO);
         var response = Response.ok().entity(result);
@@ -93,8 +106,11 @@ public class GradedCommentResource {
     @DELETE
     @Path("/{id}")
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response deleteGradedComment(@PathParam("id") Long id) {
+    public Response deleteGradedComment(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to delete GradedComment : {}", id);
+        if (!securityService.canAccess(ctx, id, GradedComment.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
         gradedCommentService.delete(id);
         var response = Response.noContent();
         HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString()).forEach(response::header);
@@ -108,7 +124,7 @@ public class GradedCommentResource {
      * @return the {@link Response} with status {@code 200 (OK)} and the list of gradedComments in body.
      */
     @GET
-    public Response getAllGradedComments(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo) {
+    public Response getAllGradedComments(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo, @Context SecurityContext ctx) {
         log.debug("REST request to get a page of GradedComments");
         var page = pageRequest.toPage();
         var sort = sortRequest.toSort();
@@ -116,10 +132,29 @@ public class GradedCommentResource {
         MultivaluedMap param = uriInfo.getQueryParameters();
         if (param.containsKey("questionId") ) {
             List questionId = (List) param.get("questionId");
+            if (!securityService.canAccess(ctx, Long.parseLong("" + questionId.get(0)), Question.class  )){
+                return Response.status(403, "Current user cannot access to this ressource").build();
+            }
             result = gradedCommentService.findGradedCommentByQuestionId(page, Long.parseLong("" + questionId.get(0)));
         }
         else{
-            result = gradedCommentService.findAll(page);
+
+            var userLogin = Optional
+            .ofNullable(ctx.getUserPrincipal().getName());
+        if (!userLogin.isPresent()){
+            throw new AccountResourceException("Current user login not found");
+        }
+        var user = User.findOneByLogin(userLogin.get());
+        if (!user.isPresent()) {
+            throw new AccountResourceException("User could not be found");
+        }
+            else if (user.get().authorities.size() >= 1 && user.get().authorities.stream().anyMatch(e1-> e1.equals(new Authority("ROLE_ADMIN")))){
+                result = gradedCommentService.findAll(page);
+
+            } else {
+                return Response.status(403, "Current user cannot access to this ressource").build();
+            }
+
 
         }
         var response = Response.ok().entity(result.content);
@@ -137,8 +172,12 @@ public class GradedCommentResource {
     @GET
     @Path("/{id}")
 
-    public Response getGradedComment(@PathParam("id") Long id) {
+    public Response getGradedComment(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to get GradedComment : {}", id);
+        if (!securityService.canAccess(ctx, id, GradedComment.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        }
+
         Optional<GradedCommentDTO> gradedCommentDTO = gradedCommentService.findOne(id);
         return ResponseUtil.wrapOrNotFound(gradedCommentDTO);
     }
